Spring WebSocket-based STOMP support
Initial work on PoC of support for STOMP based on Spring's WebSocket
support.

A message endpoint, which is a WebSocketHandler, is provided. It
handles WebSocketMessages by creating a Spring Integration message
from the payload and sending it to a channel. The WebSocketSession is
included in the message's headers.

StompInboundTransformer transforms a message containing a web socket
message payload per the STOMP specification. The command and headers
of the STOMP message are placed in the SI message's headers. The
payload of the STOMP message is used as the SI message's payload.

StompOutboundTransformer transforms an SI-representation of a STOMP
message into a byte[] ready to be sent on the wire.

WebSocketOutboundHandler sends an SI message's byte[] or String
payload as a BinaryMessage or TextMessage respectively.