Use Spring code, rather than copying and modifying
Now the Rossen's code in Spring is based on Messages and
MessageChannels we can reuse it, rather than having our own
slightly-modified versions. The only fly in the ointment is the
translation between Spring's types and Spring Integration's types.
Ideally, this would just work, but it's currently handled by some
thin adapters:

	SiToSpringMessageChannel.java
	SiToSpringMessageHandler.java
	SiToSpringSubscribableChannel.java
	SpringToSiMessageChannel.java
	SpringToSiMessageHandler.java
	SpringToSiSubscribableChannel.java