LMAX Disruptor RingBuffer-based channel
Add a new RingBufferChannel that uses a RingBuffer from the LMAX
Disruptor to store messages that are sent to the channel. Messages
are consumed from the RingBuffer on a separate thread which
dispatches them using the channel's MessageDispatcher.

This channel is somewhat similar to an ExecutorChannel in that it
provides a mechanism by which the processing of a message can be
passed from one thread to another. However, in this case, the use of a
RingBuffer allows the passing of the message from one thread to
another to be performed without locking. Also, the RingBufferChannel
only uses a single thread to consume messages from the RingBuffer.