BatchingMessageHandler that uses LMAX's RingBuffer
A BatchingMessageHandler can be used to collect messages into batches
that are of a configurable size. Each batch is released once the
batch size has been reached, or the ring buffer is empty. When a batch
is released, a message containing a list of all of the batched
message's payloads is sent to the handler's output channel.