Add XML namespace support for ring buffer channel
Ring buffer channels can now be easily configured using the XML
namespace. For example:

<ring-buffer-channel id="ringBufferChannel"/>
<ring-buffer-channel id="ringBufferChannel" size="4096"/>

A custom error handler can also be specified, along with dispatcher
configuration. The task-executor attribute of the dispatcher
configuration is not supported by a ring buffer channel. Currently,
an attempt to use it will cause a failure during parsing of the 
application context.

To support the namespace, a number of setter methods have been added
to RingBufferChannel to make it more configurable.