Implement SmartLifecycle rather than Lifecycle
Update BatchingMessageHandler and RingBufferChannel to implement
SmartLifecycle rather than Lifecycle. This ensures that their lifecycle
is managed as part of application context start and stop processing.

The start and stop methods have been improved to ensure that an already
started or stopped component is not started or stopped again.

The javadoc has been polished.